+++ sdm/libs/core/drm/hw_events_drm.cpp
@@ -349,8 +349,8 @@ void HWEventsDRM::WakeUpEventThread() {
       uint64_t exit_value = 1;
       ssize_t write_size = Sys::write_(poll_fds_[i].fd, &exit_value, sizeof(uint64_t));
       if (write_size != sizeof(uint64_t)) {
-        DLOGW("Error triggering exit fd (%d). write size = %d, error = %s", poll_fds_[i].fd,
-              write_size, strerror(errno));
+        DLOGW("Error triggering exit fd (%d). write size = %zu, error = %s", poll_fds_[i].fd,
+              static_cast<size_t>(write_size), strerror(errno));
       }
       break;
     }
@@ -673,7 +673,7 @@ void HWEventsDRM::HandlePanelDead(char *data) {
   }
 
   if (size > kMaxStringLength) {
-    DLOGE("event size %d is greater than event buffer size %zd\n", size, kMaxStringLength);
+    DLOGE("event size %d is greater than event buffer size %d\n", size, kMaxStringLength);
     return;
   }
 
@@ -723,7 +723,7 @@ void HWEventsDRM::HandleIdleTimeout(char *data) {
   }
 
   if (size > kMaxStringLength) {
-    DLOGE("event size %d is greater than event buffer size %zd\n", size, kMaxStringLength);
+    DLOGE("event size %d is greater than event buffer size %d\n", size, kMaxStringLength);
     return;
   }
 
@@ -769,7 +769,7 @@ void HWEventsDRM::HandleIdlePowerCollapse(char *data) {
   }
 
   if (size > kMaxStringLength) {
-    DLOGE("event size %d is greater than event buffer size %zd\n", size, kMaxStringLength);
+    DLOGE("event size %d is greater than event buffer size %d\n", size, kMaxStringLength);
     return;
   }
 
@@ -814,7 +814,7 @@ void HWEventsDRM::HandleHwRecovery(char *data) {
   }
 
   if (size > kMaxStringLength) {
-    DLOGE("Hardware recovery event size %d is greater than event buffer size %zd\n", size,
+    DLOGE("Hardware recovery event size %d is greater than event buffer size %d\n", size,
           kMaxStringLength);
     return;
   }
@@ -834,7 +834,7 @@ void HWEventsDRM::HandleHwRecovery(char *data) {
                                    (sizeof(event_resp->base) + sizeof(event_resp->info));
         // expect up to uint32_t from driver
         if (size_of_data > sizeof(uint32_t)) {
-          DLOGE("Size of hardware recovery event data: %" PRIu32 " exceeds %zd", size_of_data,
+          DLOGE("Size of hardware recovery event data: %zu exceeds %zu", size_of_data,
                 sizeof(uint32_t));
           return;
         }
@@ -865,7 +865,7 @@ void HWEventsDRM::HandleHistogram(char * /*data*/) {
   std::array<char, expected_size> event_data{'\0'};
   auto size = Sys::pread_(poll_fds_[histogram_index_].fd, event_data.data(), event_data.size(), 0);
   if (size != expected_size) {
-    DLOGE("event size %d is unexpected. skipping this histogram event", size);
+    DLOGE("event size %d is unexpected. skipping this histogram event", UINT32(size));
     return;
   }
